🚚 Move(all) : 절대 경로 변경 & 폴더 구조 변경
- components/common 폴더 추가
