+++ src/constants/path.js
@@ -3,8 +3,6 @@ export const SAMPLE = {
   folder: 'sample/folder',
 };
 
-// export const USERS = [{}, { userInfo: 'users/1', folders: 'users/1/folders', links: 'users/1/links' }];
-
 export const PATH = {
   user: 'users',
   folder: 'folders',
