+++ src/hooks/useGetLinks.js
@@ -1,6 +1,6 @@
 import useGetData from './useGetData';
-import { ALL_ID } from '../constants/default';
-import { PATH } from '../constants/path';
+import { ALL_ID } from 'constants/default';
+import { PATH } from 'constants/path';
 
 /**
  * @param {*} folderId 링크를 조회할 폴더의 id
