+++ src/pages/HomePage.js
@@ -2,18 +2,6 @@ import { Link } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
 import styled from 'styled-components';
 
-const HomeHead = styled.h1`
-  font-size: 3.2rem;
-`;
-
-const linkBox = {
-  margin: '10px 10px',
-  padding: '10px 10px',
-  border: 'solid pink 4px',
-  borderRadius: '20px',
-  fontSize: '24px',
-};
-
 function HomePage() {
   return (
     <>
@@ -41,3 +29,15 @@ function HomePage() {
 }
 
 export default HomePage;
+
+const HomeHead = styled.h1`
+  font-size: 3.2rem;
+`;
+
+const linkBox = {
+  margin: '10px 10px',
+  padding: '10px 10px',
+  border: 'solid pink 4px',
+  borderRadius: '20px',
+  fontSize: '24px',
+};
